+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## 1.1.3 - Developing
+* change DataForm Sample
+* fixed DataForm defaultValues problem.
 * fixed DataForm Sample async change DataForm problem.
 
 ## 1.1.2